Added a super class for bricks and coins.
Added a super-class for bricks and coins, aswell as implementing their creations independently. The reason why these should be defined in seperate is because in the future their behaviour will be different from pipes and ground.

On collision, coins and bricks might dissappear etc.

I also added that the resources used, would be disposed after the game is over/closed, in the PlayScreen-class.