Bricks are now removed when Mario's head collides with them.
With the use of bits and gdx.filter ( categoryBit and maskBits - uses the same fixture definiton as Mario) the bricks can only be collided with one time.  The bit-value gets changed from BRICK_BIT (MarioBros-clalss) to DEFAULT_BIT.

To remove the Brick texture upon collision, a method named getCell() was created. The method scaled the map to it's original size. And calculated the position of the correct tile (the tile that Mario collied with). When the correct cell is retrieved, the Brick-class sets the tile's value to "null". After is implemented the Brick texture will disappear upon collision with Marios head.